+++ apps/calculation/additional_results/chained_expressions_list_controller.h
@@ -38,8 +38,8 @@ class ChainedExpressionsListController : public ExpressionsListController {
   int numberOfRows() const override;
 
  protected:
-  int textAtIndex(char* buffer, size_t bufferSize, Escher::HighlightCell* cell,
-                  int index) override;
+  size_t textAtIndex(char* buffer, size_t bufferSize,
+                     Escher::HighlightCell* cell, int index) override;
 
  private:
   ChainableExpressionsListController* m_tail;
