+++ apps/calculation/calculation_store.cpp
@@ -327,11 +327,11 @@ char *CalculationStore::pushSerializedExpression(
     char *location, Expression e, int numberOfSignificantDigits) {
   while (true) {
     size_t availableSize = spaceForNewCalculations(location);
-    int length = availableSize > 0
-                     ? PoincareHelpers::Serialize(e, location, availableSize,
-                                                  numberOfSignificantDigits)
-                     : 0;
-    if (static_cast<size_t>(length) + 1 < availableSize) {
+    size_t length = availableSize > 0
+                        ? PoincareHelpers::Serialize(e, location, availableSize,
+                                                     numberOfSignificantDigits)
+                        : 0;
+    if (length + 1 < availableSize) {
       assert(location[length] == '\0');
       return location + length + 1;
     }
