+++ apps/inference/statistic/chi_square_and_slope/input_homogeneity_table_cell.h
@@ -45,7 +45,7 @@ class InputHomogeneityTableCell
 
  private:
   // ClearColumnHelper
-  int fillColumnName(int column, char* buffer) override;
+  size_t fillColumnName(int column, char* buffer) override;
 
   // CategoricalTableViewDataSource
   int relativeColumn(int column) const override { return column - 1; }
